requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
Django==4.2.16
django-bootstrap4==24.4
django-storages==1.14.6
djangorestframework==3.16.0
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.25.0rc1
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
idna==2.8
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
rsa==4.9.1
setuptools==69.0.3
six==1.17.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
